+++ module-apps/application-messages/windows/NewMessage.cpp
@@ -172,7 +172,7 @@ namespace gui
                                   body->getWidth() - msgStyle::recipientImg::w,
                                   msgStyle::text::h,
                                   "",
-                                  gui::Text::ExpandMode::EXPAND_NONE);
+                                  ExpandMode::EXPAND_NONE);
         recipient->setEdges(gui::RectangleEdgeFlags::GUI_RECT_EDGE_NO_EDGES);
         recipient->setInputMode(new InputMode({InputMode::phone}));
         recipient->setFont(style::window::font::mediumbold);
@@ -198,8 +198,7 @@ namespace gui
         labelMessage->setFont(style::window::font::small);
         labelMessage->setAlignment(Alignment(Alignment::ALIGN_HORIZONTAL_LEFT, Alignment::ALIGN_VERTICAL_BOTTOM));
 
-        message =
-            new gui::Text(nullptr, 0, 0, body->getWidth(), msgStyle::text::h, "", gui::Text::ExpandMode::EXPAND_UP);
+        message = new gui::Text(nullptr, 0, 0, body->getWidth(), msgStyle::text::h, "", ExpandMode::EXPAND_UP);
         message->setEdges(gui::RectangleEdgeFlags::GUI_RECT_EDGE_BOTTOM);
         message->setInputMode(new InputMode(
             {InputMode::ABC, InputMode::abc, InputMode::digit},